add middlewares cleanup index.js passport auth
