+++ server/src/routes/entryRouter.js
@@ -1,6 +1,5 @@
-
 const express = require('express');
-const {} = require('../controllers/userAPI.js');
+const verifyUser = require('../middlewares/verifyUser.js');
 const {getEntries,
        addEntry,
        updateEntry,
@@ -11,38 +10,6 @@ const {getEntries,
 const entryRouter = express.Router();
 
 // All routes should be wrapped up in try - catch blocks to prevent the entire server from crashing upon errors.
-
-//Verify user is logged in
-async function verifyUser(req, res, next) {
-    if (req.isAuthenticated() === false) {
-        return res.status(403).send("Forbidden");
-    }
-    console.log("body: " + req.body.username);
-    console.log("params: " + req.params.username);
-    if (req.session?.passport?.user.username === req.body.username)
-    {
-        console.log("username match from body");
-        next();
-    }
-    else if (req.session?.passport?.user.username !== req.body.username)
-    {
-        console.log("no username match from body");
-        next();
-    }
-    if (req.session?.passport?.user.username === req.params.username)
-    {
-        console.log("username match from url params");
-        next();
-    }
-    else if (req.session?.passport?.user.username !== req.params.username)
-    {
-        console.log("no username match from url params");
-        next();
-    }
-    //next();
-}
-
-//entryRouter.use(verifyUser);
 
 //get all entries from a user
 entryRouter.get('/get-all/:username', verifyUser, async (req, res, next) => {
@@ -63,7 +30,7 @@ entryRouter.get('/get-all/:username', verifyUser, async (req, res, next) => {
     }
 });
 
-entryRouter.put('/update', async (req, res, next) => {
+entryRouter.put('/update', verifyUser, async (req, res, next) => {
     try
     {
         let result = await updateEntry(req.body.username, req.body.entryID, req.body.info);
@@ -80,7 +47,7 @@ entryRouter.put('/update', async (req, res, next) => {
     }
 });
 
-entryRouter.post('/create', async (req, res, next) => {
+entryRouter.post('/create', verifyUser, async (req, res, next) => {
     try
     {
         let result = await addEntry(req.body.username, req.body.entry);
@@ -100,7 +67,7 @@ entryRouter.post('/create', async (req, res, next) => {
 //NOTE: Ideally the username and entryID should be in req.body.username for extra security/anonimity
 //but many servers and proxies remove the req.body from DELETE requests
 //That's why DELETE becomes a url parameter instead
-entryRouter.delete('/delete/:username/:entryID', async (req, res, next) => {
+entryRouter.delete('/delete/:username/:entryID', verifyUser, async (req, res, next) => {
     try 
     {
         let result = await deleteEntry(req.params.username, req.params.entryID);
